requirements.txt
tensorflow
numpy
flask
pillow
flask_cors